@@ -4,37 +4,37 @@
   "description": "JavaScript Game",
   "main": "main.tsx",
   "scripts": {
-    "start": "webpack-dev-server --mode development --open",
+    "start": "webpack serve --open",
     "build": "webpack"
   },
   "author": "Rinat Rezyapov",
   "license": "MIT",
   "homepage": "https://github.com/RinatRezyapov/Vault-13",
   "dependencies": {
-    "@material-ui/core": "^4.9.3",
-    "@material-ui/styles": "^4.9.0",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.11.11",
     "@types/immutability-helper": "^2.6.3",
-    "@types/node": "^13.7.0",
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "firebase": "^5.11.1",
+    "@types/node": "^13.13.52",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "file-loader": "^6.2.0",
+    "firebase": "^7.14.5",
     "fp-ts": "^1.8.1",
-    "immutability-helper": "^2.8.1",
-    "load-images": "^0.1.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "ts-loader": "^4.4.1",
-    "typescript": "^2.9.2",
-    "webpack-cli": "^3.3.11"
+    "immutability-helper": "^3.1.1",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.5"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.6.1",
+    "@types/firebase": "^3.2.1",
+    "babel-core": "^6.26.3",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "file-loader": "^2.0.0",
-    "webpack": "^4.41.6",
-    "webpack-dev-server": "^3.10.3"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^4.11.1"
   }
 }
